+++ src/makedown.js
@@ -1,10 +1,89 @@
 (function () {
+
+  let bridge = (function () {
+    const ID_KEY = '__makedown_id__'
+    const ENV_KEY = '__makedown_env__'
+    let msgId = 0
+    let callbacks = {}
+
+    function post (data, callback) {
+      msgId++
+      let message = Object.assign({
+        [ID_KEY]: msgId,
+        [ENV_KEY]: 'content-script'
+      }, data)
+      if (typeof callback === 'function') {
+        callbacks[msgId] = callback
+      }
+      window.postMessage(message, '*')
+    }
+
+    function injectScript (code) {
+      let script = document.createElement('script')
+      script.textContent = code
+      document.documentElement.appendChild(script)
+      script.parentNode.removeChild(script)
+    }
+
+    window.addEventListener('message', ({ source, data }) => {
+      let id = data[ID_KEY]
+      let env = data[ENV_KEY]
+      if (source !== window || !id || env !== 'page-script') {
+        return
+      }
+
+      if (id && callbacks[id]) {
+        delete data[ID_KEY]
+        delete data[ENV_KEY]
+        callbacks[id](data)
+        delete callbacks[id]
+      }
+    })
+
+    // add receiver in page script
+    injectScript(`
+      let makedown = {
+        dispatch (type, payload) {
+          let handler = this.actions[type]
+          if (handler) {
+            return handler(payload)
+          }
+        },
+        register (type, handler) {
+          if (typeof handler === 'function') {
+            this.actions[type] = handler
+          }
+        },
+        actions: {}
+      }
+
+      window.addEventListener('message', ({ source, data }) => {
+        let id = data['${ID_KEY}']
+        let env = data['${ENV_KEY}']
+        if (source !== window || !id || env !== 'content-script') {
+          return
+        }
+        let message = Object.assign({
+          '${ID_KEY}': id,
+          '${ENV_KEY}': 'page-script',
+        }, makedown.dispatch(data.type, data.payload))
+        window.postMessage(message, '*')
+      })`)
+
+    return {
+      post,
+      injectScript
+    }
+  })()
+
   const API_ENDPOINT = 'https://www.zhihu.com/api/v4'
+  const ZHUANLAN_ENDPOINT = 'https://zhuanlan.zhihu.com/api'
   const EXTERNAL_URL_PATTERN = /https?:\/\/link\.zhihu\.com\/\?target=/
   const VIDEO_URL_PATTERN = /https?:\/\/www\.zhihu\.com\/video\/(\d+)/
   const USER_URL_PATTERN = /https:\/\/www\.zhihu\.com\/people\/([\da-zA-Z]+)/
 
   // global states
+  let token = (document.cookie.match(/XSRF-TOKEN=([\da-zA-Z|]+)/) || [])[1]
   let editor
   let answerId
   let userData = {}
@@ -15,96 +94,120 @@
   })
   md.inline.ruler.push('inline-tex', inlineTeX)
 
-  if (!isAnswer() && !isArticle()) {
+  if (!isAnswer() && !isArticle() && location.origin !== 'https://zhuanlan.zhihu.com') {
     return
   }
 
-  if (isAnswer()) {
-    if (!window.wrappedJSObject) { // not Firefox
-      window.addEventListener('message', ({ source, data }) => {
-        if (source != window) {
-          return
-        }
-
-        if (data.type && data.type === 'makedown-answer') {
-          let { content, id } = data.payload
-          content = convertToMarkdown(content)
-          editor.value = content
-          answerId = id || null
-          syncHeight(editor)
-          focusEnd(editor)
-        } else if (data.type && data.type === 'makedown-meta') {
-          submitAnswer(data.payload)
-        }
-      }, false)
-
+  bridge.injectScript(`
+    (function () {
       /**
        * Get state from React by pretending to be React Developer Tools
        */
-      let script = document.createElement('script')
-      script.textContent = `
-        (function () {
-          function findReactComponent (elem) {
-            for (let key in elem) {
-              if (key.startsWith('__reactInternalInstance$')) {
-                return elem[key]._currentElement._owner._instance
-              }
-            }
-            return null
+      function findReactComponent (elem) {
+        for (let key in elem) {
+          if (key.startsWith('__reactInternalInstance$')) {
+            return elem[key]._currentElement._owner._instance
           }
+        }
+        return null
+      }
 
-          window.addEventListener('message', ({ source, data }) => {
-            if (data.type === 'makedown-query') {
-              let node = document.querySelector('[data-makedown-answer]').parentNode.parentNode
-              let instance = findReactComponent(node)
-              if (!instance) {
-                return
-              }
+      makedown.register('query-answer', () => {
+        let node = document.querySelector('[data-makedown-answer]').parentNode.parentNode
+        let instance = findReactComponent(node)
+        if (!instance) {
+          return
+        }
+        let content = instance.props.defaultValue || ''
+        return { content }
+      })
 
-              if (!data.payload || !data.payload.meta) {
-                let { answer } = instance.props
-                content = answer ? answer.editableContent : ''
-                id = answer ? answer.id : null
-                window.postMessage({
-                  type: 'makedown-answer',
-                  payload: { content, id }
-                }, '*')
-              } else if (data.payload && data.payload.meta) {
-                let node = document.querySelector('[data-makedown-answer]').parentNode
-                let instance = findReactComponent(node)
-                if (!instance) {
-                  return
-                }
-                let { commentPermission, reshipmentSettings } = instance.state
-                window.postMessage({
-                  type: 'makedown-meta',
-                  payload: {
-                    reshipmentSettings,
-                    commentPermission
-                  }
-                }, '*')
-              }
-            }
-          })
-        })()`
-      document.documentElement.appendChild(script)
-      script.parentNode.removeChild(script)
-    }
-  }
+      makedown.register('query-meta', () => {
+        let node = document.querySelector('[data-makedown-answer]').parentNode
+        let instance = findReactComponent(node)
+        if (!instance) {
+          return
+        }
+        let { commentPermission, reshipmentSettings } = instance.state
+        return {
+          reshipmentSettings,
+          commentPermission
+        }
+      })
+
+      makedown.register('query-article', () => {
+        let node = document.querySelector('.PublishPanel-wrapper')
+        let instance = findReactComponent(node)
+        if (!instance) {
+          return
+        }
+        let { draft = {} } = instance.props
+        return { content: draft.content || '' }
+      })
+
+      makedown.register('sync-article', ({ content }) => {
+        let node = document.querySelector('.PublishPanel-wrapper')
+        let instance = findReactComponent(node)
+        if (!instance) {
+          return
+        }
+        let { draft } = instance.props
+        draft.content = content
+        instance.setState({
+          disabled: content.length < 10
+        })
+      })
+
+      // makedown.register('next-step', function ({ content }) {
+      //   let node = document.querySelector('.PublishPanel-wrapper')
+      //   let instance = findReactComponent(node)
+      //   if (!instance) {
+      //     return
+      //   }
+      //   let { draft, publishColumn, columns } = instance.props
+      //   draft.content = content
+      //   let { state, isTitleImageFullScreen, id } = draft
+      //   let { searchItems, commentPermissionVal, selectInputVal, step } = instance.state
+      //   let redirect = () => {
+      //       window.location.href = '/p/' + id
+      //   }
+
+      //   if (state === 'published') {
+      //     publishColumn(draft, searchItems, state, isTitleImageFullScreen).then(redirect);
+      //   } else if (state === 'draft') {
+      //     if (step === 2) {
+      //       if (columns && columns.length) {
+      //         instance.setState({ step: 3 })
+      //       } else {
+      //         publishColumn(draft, searchItems, state, isTitleImageFullScreen, commentPermissionVal).then(redirect)
+      //       }
+      //     } else {
+      //       let column = selectInputVal > 0 ? columns[selectInputVal - 1] : null
+      //       publishColumn(draft, searchItems, state, isTitleImageFullScreen, commentPermissionVal, column).then(redirect)
+      //     }
+      //   }
+      // })
+    })()`)
 
   $(function () {
     if (document.querySelector('.AnswerForm-editor')) {
       makedownAnswer()
+      return
+    } else if (document.querySelector('.PostEditor')) {
+      makedownArticleReact()
       return
     }
 
     let observer = new MutationObserver(mutations => {
       mutations.forEach(({ type, target, addedNodes }) => {
         if (type === 'childList') {
-          if (addedNodes.length && [...addedNodes].find(isAnswerEditor)) {
+          let added = [...addedNodes]
+          if (added.find(isAnswerEditor)) {
             makedownAnswer()
           } else if (target.matches('.entry-content')) {
             makedownArticle()
+          } else if (added.find(isArticleEditor)) {
+            makedownArticleReact()
           }
         }
       })
@@ -134,8 +237,8 @@
       return
     }
 
-    document.body.classList.add('makedown-enabled')
-    document.body.classList.add('makedown-answer')
+    addClass('makedown-enabled')
+    addClass('makedown-answer')
 
     editor = createEditor(richEditor, {
       placeholder: '写回答...'
@@ -172,8 +275,8 @@
       return
     }
 
-    document.body.classList.add('makedown-enabled')
-    document.body.classList.add('makedown-article')
+    addClass('makedown-enabled')
+    addClass('makedown-article')
 
     let placeholder = '请输入正文'
     editor = createEditor(richEditor, {
@@ -198,6 +301,54 @@
       editor.value = convertToMarkdown(content.innerHTML)
     }
     syncHeight(editor)
+  }
+
+  function makedownArticleReact () {
+    if (document.getElementById('makedown')) {
+      return
+    }
+
+    /**
+     * Initialize the editor overlay
+     */
+    let richEditor = document.querySelector('.PostEditor')
+    if (!richEditor) {
+      return
+    }
+
+    addClass('makedown-enabled')
+    addClass('makedown-article')
+    addClass('makedown-article-react')
+
+    let placeholder = '请输入正文'
+    editor = createEditor(richEditor, {
+      placeholder
+    })
+
+    editor.addEventListener('focus', () => { editor.placeholder = '' })
+    editor.addEventListener('blur', () => { editor.placeholder = placeholder })
+
+    let sync = debounce(() => {
+      saveArticleDraft(getArticleId(), convertToHTML(editor.value))
+    }, 1000)
+    editor.addEventListener('input', sync)
+
+    /**
+     * Hijacking DOM events
+     */
+    // document.body.addEventListener('click', e => {
+    //   let submitButton = document.querySelector('.PublishPanel-wrapper .publish-button')
+    //   if (submitButton && submitButton.contains(e.target)) {
+    //     handleNextStep()
+    //     e.stopPropagation()
+    //   }
+    // }, true)
+
+    /**
+     * Get editable content from page script
+     */
+    setTimeout(queryArticle, 666) // oops, don't know how to handle callback
+                                  // wait for a little while
   }
 
   function createEditor (container, options = {}) {
@@ -298,7 +449,6 @@
     }
     let data = new FormData()
     data.append(field, file)
-    let token = (document.cookie.match(/XSRF-TOKEN=([\da-zA-Z|]+)/) || [])[1]
     let headers = {}
     if (token) {
       headers['X-XSRF-TOKEN'] = token
@@ -333,7 +483,7 @@
       return
     }
 
-    queryMeta();
+    queryMetaAndSubmit();
 
     e.preventDefault()
     e.stopPropagation()
@@ -361,24 +511,23 @@
       }
       let { answer } = instance.props
       if (answer) {
-        content = convertToMarkdown(answer.editableContent)
-        answerId = answer.id
-        editor.value = content
+        editor.value = convertToMarkdown(answer.editableContent)
       } else {
-        content = ''
-        answerId = null
+        editor.value = ''
       }
       syncHeight(editor)
       focusEnd(editor)
     } else {
       $('.AnswerForm-editor').attr('data-makedown-answer', '')
-      window.postMessage({
-        type: 'makedown-query'
-      }, '*')
+      bridge.post({ type: 'query-answer' }, ({ content }) => {
+        editor.value = convertToMarkdown(content)
+        syncHeight(editor)
+        focusEnd(editor)
+      })
     }
   }
 
-  function queryMeta (submit) {
+  function queryMetaAndSubmit () {
     if (window.wrappedJSObject) {
       let doc = window.wrappedJSObject.document
       let node = doc.querySelector('.AnswerForm-editor').parentNode
@@ -390,14 +539,114 @@
       submitAnswer({ commentPermission, reshipmentSettings })
     } else {
       $('.AnswerForm-editor').attr('data-makedown-answer', '')
-      window.postMessage({
-        type: 'makedown-query',
-        payload: {
-          meta: true
-        }
-      }, '*')
+      bridge.post({ type: 'query-meta' }, ({ commentPermission, reshipmentSettings }) => {
+        submitAnswer({ commentPermission, reshipmentSettings })
+      })
     }
   }
+
+  function queryArticle () {
+    if (window.wrappedJSObject) {
+      let doc = window.wrappedJSObject.document
+      let node = doc.querySelector('.PublishPanel-wrapper')
+      let instance = findReactComponent(node)
+      if (!instance) {
+        return
+      }
+      let { draft = {} } = instance.props
+      editor.value = convertToMarkdown(draft.content || '')
+      syncHeight(editor)
+    } else {
+      bridge.post({ type: 'query-article' }, ({ content }) => {
+        editor.value = convertToMarkdown(content)
+        syncHeight(editor)
+      })
+    }
+  }
+
+  function saveArticleDraft (articleId, content) {
+    if (!content) {
+      return
+    }
+    syncArticleReact()
+    let url =`${ZHUANLAN_ENDPOINT}/drafts/${articleId || ''}`
+    let saving = fetch(url, {
+        method: articleId ? 'PATCH' : 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'X-XSRF-TOKEN': token
+        },
+        credentials: 'include',
+        body: JSON.stringify({
+          content
+        })
+      })
+
+    if (!articleId) {
+      saving
+        .then(json)
+        .then(({ id }) => {
+          if (!articleId) {
+            history.pushState({}, document.title, `p/${id}/edit`)
+          }
+        })
+    }
+  }
+
+  function syncArticleReact () {
+    let content = convertToHTML(editor.value)
+    if (window.wrappedJSObject) {
+      let doc = window.wrappedJSObject.document
+      let node = doc.querySelector('.PublishPanel-wrapper')
+      let instance = findReactComponent(node)
+      if (!instance) {
+        return
+      }
+      let { draft } = instance.props
+      draft.content = content
+      instance.setState({
+        disabled: content.length < 10
+      })
+    } else {
+      bridge.post({ type: 'sync-article', payload: { content }})
+    }
+  }
+
+  // function handleNextStep () {
+  //   let content = convertToHTML(editor.value)
+  //   if (window.wrappedJSObject) {
+  //     let doc = window.wrappedJSObject.document
+  //     let node = doc.querySelector('.PublishPanel-wrapper')
+  //     let instance = findReactComponent(node)
+  //     if (!instance) {
+  //       return
+  //     }
+  //     let { draft, publishColumn, columns } = instance.props
+  //     draft.content = content
+  //     let { state, isTitleImageFullScreen, id } = draft
+  //     let { searchItems, commentPermissionVal, selectInputVal, step } = instance.state
+  //     let redirect = () => {
+  //         window.location.href = `/p/${id}`
+  //     }
+
+  //     if (state === 'published') {
+  //       publishColumn(draft, searchItems, state, isTitleImageFullScreen).then(redirect);
+  //     } else if (state === 'draft') {
+  //       if (step === 2) {
+  //         if (columns && columns.length) {
+  //           instance.setState({ step: 3 })
+  //         } else {
+  //           publishColumn(draft, searchItems, state, isTitleImageFullScreen, commentPermissionVal).then(redirect)
+  //         }
+  //       } else {
+  //         let column = selectInputVal > 0 ? columns[selectInputVal - 1] : null
+  //         publishColumn(draft, searchItems, state, isTitleImageFullScreen, commentPermissionVal, column).then(redirect)
+  //       }
+  //     }
+  //   } else {
+  //     bridge.post({ type: 'next-step', payload: { content }})
+  //   }
+  // }
 
   function focusEnd (editor) {
     setTimeout(() => {
@@ -625,8 +874,16 @@
     return location.host === 'zhuanlan.zhihu.com' && location.pathname.match(/^\/(?:p\/\d+\/edit|write$)/)
   }
 
+  function getArticleId () {
+    return (location.host === 'zhuanlan.zhihu.com' && location.pathname.match(/^\/p\/(\d+)/) || [])[1]
+  }
+
   function isAnswerEditor (node) {
     return node.nodeType === Node.ELEMENT_NODE && node.matches('.AnswerForm-editor')
+  }
+
+  function isArticleEditor (node) {
+    return node.nodeType === Node.ELEMENT_NODE && node.matches('.PostEditor')
   }
 
   function findReactComponent (elem) {
@@ -636,6 +893,10 @@
       }
     }
     return null
+  }
+
+  function addClass (name) {
+    document.documentElement.classList.add(name)
   }
 
   function inlineTeX (state, silent) {
@@ -686,5 +947,24 @@
     }
     state.pos += marker.length
     return true
+  }
+
+  // https://davidwalsh.name/javascript-debounce-function
+  function debounce (func, wait, immediate) {
+    let timeout
+    return function (...args) {
+      let later = () => {
+        timeout = null
+        if (!immediate) {
+          func.apply(this, args)
+        }
+      }
+      let callNow = immediate && !timeout
+      clearTimeout(timeout)
+      timeout = setTimeout(later, wait)
+      if (callNow) {
+        func.apply(this, args)
+      }
+    }
   }
 })()