+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+0.1.2
+* Fix answer edit.
+
+0.1.1
+* Fix article edit.
+
+0.1.0
+* Fix article support for both Angular & React version.
+
+0.0.7
+* Fix answer posting due to Zhihu's latest updates.
+* Not working for aritcles right now.
+
 0.0.6
 * Make existing inline video compatible.